+++ src/CuiLib/Checkers/ValueChecker.Implementation.cs
@@ -1,4 +1,4 @@
-﻿using CuiLib.Internal;
+﻿using CuiLib.Internal.Versions;
 using System;
 using System.Collections;
 using System.Collections.Generic;
