+++ src/CuiLib/Commands/CommandCollection.cs
@@ -1,4 +1,5 @@
 ﻿using CuiLib.Internal;
+using CuiLib.Internal.Versions;
 using System;
 using System.Collections;
 using System.Collections.Generic;
